# siret/public/api.yaml
openapi: 3.0.3
info:
  title: Entreprise API
  description: Here goes description
  version: "0.1"
servers:
  - url: http://localhost:8000/api
    description: local server
  - url: http://127.0.0.1:8000/api
    description: local server
paths:
  /entreprises:
    get:
      summary: Liste des établissements dans la base de données
      parameters:
        - in: query
          name: query
          schema:
            type: string
          description: Valeur à rechercher dans la base de données
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: L'index de la page en cours en fonction du nombre de résultat et de la limite
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
          description: Le nombre d'élements retournés. (6 par défaut)
      responses:
        '200': 
          description: Un tableau JSON avec les informations des entreprises
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: object
                  properties:
                    statut:
                      type: integer
                      format: int64
                      example: 200
                      description: 200 OK
                    datas:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: integer
                            description: Id dans la base de données
                            example: 4
                          siret:
                            type: integer
                            description: Numéro SIRET (14 chiffres)
                            example: 85106293500016
                          siren:
                            type: integer
                            description: Numéro SIREN (9 chiffres)
                            example: 851062935
                          tva:
                            type: string
                            description: Numéro TVA (FR + clé TVA + SIREN)
                            example: "FR19851062935"
                          nom:
                            type: string
                            description: Nom de l'entreprise
                            example: "SCEA DE L'ABBAYE D'OIGNY"
                          numeroVoie:
                            type: integer
                            description: Numéro de la voie
                            example: null
                          typeVoie:
                            type: string
                            description: Type de voie (Rue, Avenue, ...)
                            example: null
                          libelleVoie:
                            type: string
                            description: Nom de la voie
                            example: "ABBAYE D'OIGNY"
                          codePostal:
                            type: integer
                            description: Code Postal
                            example: 21450
                          libelleCommune:
                            type: string
                            description: Nom de la commune
                            example: "OIGNY"
                          dateCreation:
                            type: string
                            description: Date de la création de l'entreprise
                            example: "2019-06-01"
                          updated_at:
                            type: string
                            description: Date de la dernière mise à jour de l'entrée
                            example: "2022-12-30T12:48:03.000000Z"
                          created_at:
                            type: string
                            description: Date de la création de l'entrée
                            example: "2022-12-30T10:36:35.000000Z"
    post:
      summary: Ajout d'un établissement dans la base de données
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                siret: 
                  description: Numéro SIRET de l'entreprise (14 chiffres)
                  type: integer
                siren:
                  description: Numéro SIREN (9 chiffres)
                  type: integer
                tva:
                  description: Numéro TVA (FR + clé TVA + SIREN)
                  type: string
                nom:
                  description: Nom de l'entreprise
                  type: string
                numeroVoie:
                  description: Numéro de la voie
                  type: integer
                typeVoie:
                  description: Type de voie (Rue, Avenue, ...)
                  type: string
                libelleVoie:
                  description: Nom de la voie
                  type: string
                codePostal:
                  description: Code Postal
                  type: integer
                libelleCommune:
                  description: Nom de la commune
                  type: string
                dateCreation:
                  description: Date de la création de l'entreprise
                  type: string
              required: [siret, siren, tva, nom, libelleVoie, codePostal, libelleCommune, dateCreation]
      responses:
        '201': 
          description: Un tableau JSON avec les informations de l'entreprise
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: object
                  properties:
                    statut:
                      type: integer
                      format: int64
                      example: 201
                      description: 201 Created
                    datas:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: Id de l'entreprise
                          example: 4
                        siret:
                          type: integer
                          description: Numéro SIRET (14 chiffres)
                          example: 85106293500016
                        siren:
                          type: integer
                          description: Numéro SIREN (9 chiffres)
                          example: 851062935
                        tva:
                          type: string
                          description: Numéro TVA (FR + clé TVA + SIREN)
                          example: "FR19851062935"
                        nom:
                          type: string
                          description: Nom de l'entreprise
                          example: "SCEA DE L'ABBAYE D'OIGNY"
                        numeroVoie:
                          type: integer
                          description: Numéro de la voie
                          example: null
                        typeVoie:
                          type: string
                          description: Type de voie (Rue, Avenue, ...)
                          example: null
                        libelleVoie:
                          type: string
                          description: Nom de la voie
                          example: "ABBAYE D'OIGNY"
                        codePostal:
                          type: integer
                          description: Code Postal
                          example: 21450
                        libelleCommune:
                          type: string
                          description: Nom de la commune
                          example: "OIGNY"
                        dateCreation:
                          type: string
                          description: Date de la création de l'entreprise
                          example: "2019-06-01"
                        updated_at:
                          type: string
                          description: Date de la dernière mise à jour de l'entrée
                          example: "2022-12-30T12:48:03.000000Z"
                        created_at:
                          type: string
                          description: Date de la création de l'entrée
                          example: "2022-12-30T10:36:35.000000Z"
        '400':
          description: Erreur 400, impossible d'ajouter l'entrée dans la base de données
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: object
                  properties:
                    statut:
                      type: integer
                      format: int64
                      example: 400
                      description: 400 Bad Request
                    error:
                      type: string
                      example: STR_DATABASE_INSERT_ERROR
                      description: Description de l'erreur
...